We need to properly set the gateway.public.host parameter during server setup. For this to happen, we need to know the DNS name of the load balancer. To handle the scenario where the user doesn't assing a specific DNS name to the load balancer, we need to create the load balancer first.
Since there's no way in CloudFormation to attach an instance to a load
balancer after it's created, I have the primary use a small python
script to do it for us. This means we have to pass in the load balancer
name, along with the desired DNS name.

Also, fixed some labels and descriptions.